handle ssm agent start

+++ src/ecs-additions/provision.sh
@@ -3,25 +3,59 @@
 set -e
 set -x
 
+export OS=`uname -r`
+
 # Expected environment variables
 #   EBS_AUTOSCALE_VERSION
 #   EBS_AUTOSCALE_FILESYSTEM
 #   ARTIFACT_ROOT_URL
 #   ARTIFACT_HTTP_ROOT_URL
 #   ARTIFACT_S3_ROOT_URL
-#   WORKFLOW_ORCHESTRATOR
+#   WORKFLOW_ORCHESTRATOR (OPTIONAL)
+
+printenv
+
+# start ssm-agent
+if [[ $OS =~ "amzn1" ]]; then
+    start amazon-ssm-agent
+elif [[ $OS =~ "amzn2" ]]; then
+    systemctl enable amazon-ssm-agent
+    systemctl start amazon-ssm-agent
+else
+    echo "unsupported os: $os"
+    exit 100
+fi
 
 function ecs() {
-    case $1 in
-        disable)
-            systemctl stop ecs
-            systemctl stop docker
-            ;;
-        enable)
-            systemctl start docker
-            systemctl enable --now --no-block ecs  # see: https://github.com/aws/amazon-ecs-agent/issues/1707
-            ;;
-    esac
+    
+    if [[ $OS =~ "amzn1" ]]; then
+        # Amazon Linux 1 uses upstart for init
+        case $1 in
+            disable)
+                stop ecs
+                service docker stop
+                ;;
+            enable)
+                service docker start
+                start ecs
+                ;;
+        esac
+    elif [[ $OS =~ "amzn2" ]]; then
+        # Amazon Linux 2 uses systemd for init
+        case $1 in
+            disable)
+                systemctl stop ecs
+                systemctl stop docker
+                ;;
+            enable)
+                systemctl start docker
+                systemctl enable --now --no-block ecs  # see: https://github.com/aws/amazon-ecs-agent/issues/1707
+                ;;
+        esac
+    else
+        echo "unsupported os: $os"
+        exit 100
+    fi
 }
 
 # make sure that docker and ecs are running on script exit to avoid
